BACK-1169 Sync endpoint does not wait end of sync. It returns true only when synced status
The endpoint
/{{ url  }}/pools/{{ pool_name }}/wallets/{{ walletname  }}/accounts/{{ accidx  }}/operations/synchronize
uses the sync manager. The sync manager can tell "sync result true" the synchronizer is in synced status.

If there is no current sync, it triggers a sync.

This endpoint should return the status of the current synchronizer instead of a Seq (?!) of SynchronizationResult.
( At this moment, Vault does not use the payload )